refactor(auth): split api/endpoints/auth.py into an 11-module package
Pure mechanical extraction for #284 A3.5 — no behaviour change. All 54
top-level defs move verbatim (verified AST-identical against HEAD) out of the
2817-line module into app/api/endpoints/auth/:

  dependencies     get_current_* + oauth2_scheme (route-free, so the ~30
                   importers never pull a router into their import graph)
  authenticators   per-auth_type credential verification
  login            /token + /login and its orchestration
  registration     /register, /password-policy, /password-reset/*
  profile          /me, /session, /me/certificate
  keycloak, pki    external IdP flows
  methods          /methods, /banner
  mfa_tokens       MFA token minting + single-use replay protection
  mfa              the MFA endpoints
  sessions         refresh rotation, logout, active sessions

Each endpoint module owns an APIRouter that __init__.py includes in the
original declaration order, and __init__ re-exports every name the flat module
exported so `from app.api.endpoints.auth import get_current_active_user` (~30
call sites) is unchanged.

The full app route table is byte-identical to master (409 routes: same paths,
methods, and endpoint names, same /api/auth registration order) — that is what
proves the router recomposition is behaviour-preserving.

Test call sites that patched or inspected the flat module now target the module
that owns the symbol (test_mfa_security, test_mfa_integration,
test_fedramp_controls, test_handler_and_ws_hardening,
test_mfa_replay_and_registration); inspect.getsource() on a package would only
see __init__.py, so the source-scanning assertions now scan every module.
